Added tests for user mutations
