+++ client/pages/home.tsx
@@ -10,42 +10,42 @@ import { getRequest } from '../lib/fetch';
 import { iPost, iTag } from '../@types';
 
 interface Props {
-    posts: iPost[];
-    tags: { tag: iTag; post: iPost }[];
+	posts: iPost[];
+	tags: { tag: iTag; post: iPost }[];
 }
 
 const Feed: NextPage<Props> = ({ posts, tags }) => {
-    return (
-        <>
-            <Grid container>
-                <Grid item sm={3}>
-                    <TagBarLayout>
-                        {tags.map((tag) => (
-                            <TagCard tag={tag} />
-                        ))}
-                    </TagBarLayout>
-                </Grid>
+	return (
+		<>
+			<Grid container>
+				<Grid item sm={3}>
+					<TagBarLayout>
+						{tags.map((tag) => (
+							<TagCard tag={tag} />
+						))}
+					</TagBarLayout>
+				</Grid>
 
-                <Grid item sm={9}>
-                    <ChatLayout>
-                        <Grid container direction="column">
-                            {posts.map((post, i) => (
-                                <MessagePost key={post._id} out={i === 48} post={post} />
-                            ))}
-                        </Grid>
-                    </ChatLayout>
-                </Grid>
-            </Grid>
-        </>
-    );
+				<Grid item sm={9}>
+					<ChatLayout>
+						<Grid container direction="column">
+							{posts.map((post, i) => (
+								<MessagePost key={post._id} out={i === 48} post={post} />
+							))}
+						</Grid>
+					</ChatLayout>
+				</Grid>
+			</Grid>
+		</>
+	);
 };
 
 export const getServerSideProps = async (ctx: NextPageContext) => {
-    authInitialProps(true)(ctx);
-    const posts = await getRequest('/posts').then((res) => res.json());
-    const tags = await getRequest('/tags').then((res) => res.json());
+	authInitialProps(true)(ctx);
+	const posts = await getRequest('/posts').then((res) => res.json());
+	const tags = await getRequest('/tags').then((res) => res.json());
 
-    return { props: { posts, tags } };
+	return { props: { posts, tags } };
 };
 
 export default Feed;
